Fixed bug in spGisModelScadaData stored procedure.

+++ WbEasyCalc/WbEasyCalc/DataRepository/DataRepository.cs
@@ -116,7 +116,7 @@ namespace DataRepository
                 p.Add("@UnauthCons_MeterTampBypEtcEstNo_D14", dbType: DbType.Double, direction: ParameterDirection.Output);
                 p.Add("@UnauthCons_MeterTampBypEtcErrorMargin_F14", dbType: DbType.Double, direction: ParameterDirection.Output);
                 p.Add("@UnauthCons_MeterTampBypEtcConsLitPerCustDay_J14", dbType: DbType.Double, direction: ParameterDirection.Output);
-                p.Add("@MetErrors_DetailedManualSpec_J6", dbType: DbType.Double, direction: ParameterDirection.Output);                      //
+                p.Add("@MetErrors_DetailedManualSpec_J6", dbType: DbType.Int32, direction: ParameterDirection.Output);                      //
                 p.Add("@MetErrors_BilledMetConsWoBulkSupMetUndrreg_H8", dbType: DbType.Double, direction: ParameterDirection.Output);
                 p.Add("@MetErrors_BilledMetConsWoBulkSupErrorMargin_N8", dbType: DbType.Double, direction: ParameterDirection.Output);
